refactor(x402): split payment submission out of x402-client
x402-client.ts mixed request orchestration with the payment-building
and signing logic (simulate the SEP-41 transfer, derive the signature
expiration, verify and sign the auth entry). Extract that into a new
x402-payment module so x402-client.ts is left as the orchestration
layer: guard/budget checks, the fetch/retry flow, and settlement
recording.

The pure discovery layer (decode/select a 402 challenge) already lived
in x402-guards.ts and is unchanged. x402-client.ts still re-exports
both, so its published API is unaffected.

Adds module-level tests for the new x402-payment module and moves the
expirationOffsetFor derivation tests there, next to the function.
